split up date controls (#200)
